refactor(web): extract shared client logic to core.js; slim app.js files
New file: client-samples/web/App/core.js (~580 lines)
  Exports parameterised versions of every action and render helper:
  createBaseModel(), renderBase(), connect(), disconnect(), startAudio(),
  stopAudio(), startCamera(), stopCamera(), sendPing(), sendCustom(),
  enumerateCameras(), wireBaseEvents(). No module-level state — callers
  own their model object.

  Already incorporates:
  - VOICE_PROCESSING as the default audio mode
  - Camera enumeration on CONNECTED (populates selector before first start)
  - Camera stop on RECONNECTING (server/client state sync on reconnect)
  - Camera selector shown for >= 1 devices

New symlink: client-samples/web-xr/App/core.js → ../../web/App/core.js
  Single source of truth; web-xr gets core updates automatically.

client-samples/web/App/app.js  (~80 lines, was 277)
  Now just imports from core.js, owns the model instance, error toast,
  and bootstrap. All duplicated logic removed.

client-samples/web-xr/App/app.js  (~324 lines, was 700+)
  Imports base logic from core.js. Extends only with XR-specific fields
  (xrHost, xrPort, xrState, xrError), render additions (XR Stream panel),
  cert verification, CloudXRStream lifecycle, and onDataReceived intercept.

Also fixed stale docstring in core.js that said web/App/app.js does not
import from it (it does now).
