+++ client-samples/web-xr/App/core.js
@@ -1,0 +1,1 @@
+../../web/App/core.js
